*фикс выравнивания памяти

+++ jni/imageloader.h
@@ -38,25 +38,25 @@ private:
 #pragma pack(2)
 
 struct ImageType_BMP_FileHeader {
-  unsigned short  bfType;
-  unsigned long   bfSize;
-  unsigned short  bfReserved1;
-  unsigned short  bfReserved2;
-  unsigned long   bfOffBits;
+  uint16_t        bfType;
+  uint32_t        bfSize;
+  uint16_t        bfReserved1;
+  uint16_t        bfReserved2;
+  uint32_t        bfOffBits;
 };
 
 struct ImageType_BMP_InfoHeader {
-  unsigned long   biSize;
-  long            biWidth;
-  long            biHeight;
-  unsigned short  biPlanes;
-  unsigned short  biBitCount;
-  unsigned long   biCompression;
-  unsigned long   biSizeImage;
-  long            biXPelsPerMeter;
-  long            biYPelsPerMeter;
-  unsigned long   biClrUsed;
-  unsigned long   biClrImportant;
+  uint32_t        biSize;
+  int32_t         biWidth;
+  int32_t         biHeight;
+  uint16_t        biPlanes;
+  uint16_t        biBitCount;
+  uint32_t        biCompression;
+  uint32_t        biSizeImage;
+  int32_t         biXPelsPerMeter;
+  int32_t         biYPelsPerMeter;
+  uint32_t        biClrUsed;
+  uint32_t        biClrImportant;
 };
 
 #pragma pack()
